third_party/WebKit/LayoutTests/svg/animations/animate-elem-15-t-drt-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests


A copy of the corresponding W3C-SVG-1.1 test, which dumps the animation at certain times

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.width.animVal.value is 300
PASS rect.width.baseVal.value is 300
PASS rect.width.animVal.value is 232.5
PASS rect.width.baseVal.value is 300
PASS rect.width.animVal.value is 232.5
PASS rect.width.baseVal.value is 300
PASS rect.width.animVal.value is 165
PASS rect.width.baseVal.value is 300
PASS rect.width.animVal.value is 165
PASS rect.width.baseVal.value is 300
PASS rect.width.animVal.value is 30
PASS rect.width.baseVal.value is 300
PASS rect.width.animVal.value is 30
PASS rect.width.baseVal.value is 300
PASS rect.width.animVal.value is 30
PASS rect.width.baseVal.value is 300
PASS successfullyParsed is true

TEST COMPLETE
